refactor per page styles
