--- assets/public/assets/styles/ui/home/main-scene.module.scss
@@ -1,180 +1,0 @@
-@use "@styles/mixins.module.scss";
-
-.mc-main-scene-container {
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    height: min-content;
-    position: relative;
-
-    img {
-        @include mixins.mc-absolute;
-        // border: 1px solid orangered;
-        pointer-events: none;
-    }
-}
-
-.mc-main-scene-background-light {
-    // bottom: 0;
-    background-color: #fff;
-    border-radius: 319px;
-    filter: blur(200px);
-    height: 637px;
-    // left: 4%;
-    mix-blend-mode: overlay;
-    // position: absolute;
-    // top: 0;
-    width: 637px;
-}
-
-.mc-main-scene-box-img {
-    filter: brightness(1.3) drop-shadow(2px 4px 6px black);
-    max-height: 350px;
-    image-rendering: -webkit-optimize-contrast;
-    position: absolute;
-    transform: translate(0, -9%);
-    max-width: 565px;
-    z-index: 5;
-
-    @include mixins.keyframes(box-img) {
-        0% {
-            transform: translate(0, -9%);
-
-        }
-        25% {
-            transform: translate(2.5%, -10.5%);
-
-        }
-        50% {
-            transform: translate(4%, -9%);
-
-        }
-        75% {
-            transform: translate(1.5%, -7.5%);
-
-        }
-        100% {
-            transform: translate(0, -9%);
-
-        }
-    };
-    @include mixins.animate(box-img, 6s, linear, infinite);
-}
-
-.mc-main-scene-left-ghost {
-    filter: brightness(1.3) drop-shadow(2px 4px 6px black);
-    image-rendering: -webkit-optimize-contrast;
-    max-height: 408px;
-    max-width: 232px;
-    transform: translate(-125%, -15%);
-    z-index: 10;
-
-    @include mixins.keyframes(left-ghost) {
-        0% {
-            transform: translate(-125%, -15%);
-        }
-        25% {
-            transform: translate(-124%, -16.5%);
-        }
-        50% {
-            transform: translate(-126%, -14%);
-        }
-        75% {
-            transform: translate(-127%, -12%);
-        }
-        100% {
-            transform: translate(-125%, -15%);
-        }
-    };
-    @include mixins.animate(left-ghost, 4s, linear, infinite);
-}
-
-.mc-main-scene-back-ghost {
-    filter: brightness(0.85) contrast(1.05);
-    image-rendering: -webkit-optimize-contrast;
-    max-height: 279px;
-    max-width: 214px;
-    transform: translate(7%, -70%);
-    z-index: 0;
-
-    @include mixins.keyframes(back-ghost) {
-        0% {
-            transform: translate(7%, -70%);
-        }
-        25% {
-            transform: translate(6%, -69%);
-        }
-        50% {
-            transform: translate(8%, -68%);
-        }
-        75% {
-            transform: translate(6%, -69%);
-        }
-        100% {
-            transform: translate(7%, -70%);
-        }
-    };
-    @include mixins.animate(back-ghost, 3s, linear, infinite);
-}
-
-.mc-main-scene-right-ghost {
-    filter: brightness(1.3) drop-shadow(2px 4px 6px black);
-    image-rendering: -webkit-optimize-contrast;
-    max-height: 240px;
-    max-width: 186px;
-    transform: translate(101%, 35%);
-    z-index: 9;
-
-    @include mixins.keyframes(right-ghost) {
-        0% {
-            transform: translate(101%, 35%);
-        }
-        25% {
-            transform: translate(100%, 36%);
-        }
-        50% {
-            transform: translate(102%, 34%);
-        }
-        75% {
-            transform: translate(103%, 36%);
-        }
-        100% {
-            transform: translate(101%, 35%);
-        }
-    };
-    @include mixins.animate(right-ghost, 5s, linear, infinite);
-}
-
-// .mc-main-scene-container img,
-.mc-main-scene-container .mc-main-scene-star {
-    @include mixins.mc-absolute;
-
-    left: 0;
-    pointer-events: none;
-    top: 0;
-}
-
-.mc-main-scene-star path:nth-child(1) {
-    transform: translate(15%, 87%);
-}
-
-.mc-main-scene-star path:nth-child(2) {
-    transform: translate(45%, -15%);
-}
-
-.mc-main-scene-star path:nth-child(3) {
-    transform: translate(29%, 17%);
-}
-
-.mc-main-scene-star path:nth-child(4) {
-    transform: translate(45%, 80%);
-}
-
-.mc-main-scene-star path:nth-child(5) {
-    transform: translate(70%);
-}
-
-.mc-main-scene-star {
-    overflow: visible;
-    z-index: 10;
-}
