--- assets/public/assets/styles/ui/home/main.module.scss
@@ -1,24 +1,0 @@
-.mc-home-text-container {
-    align-items: flex-start;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding-top: 16px;
-
-    & h1 {
-        font-family: Pangram;
-        font-size: 72px;
-        font-weight: 700;
-        line-height: 74p;
-        letter-spacing: 0.4px;
-        text-transform: capitalize;
-    }
-    
-    & p {
-        font-family: Pangram;
-        font-size: 24px;
-        font-weight: normal;
-        line-height: 36px;
-        letter-spacing: 0.3px;
-    }
-}
