--- assets/public/assets/styles/ui/home/requirements.module.scss
@@ -1,38 +1,0 @@
-.mc-requirements-icon-container {
-    margin-bottom: 48px;
-
-    & svg {
-        height: 160px;
-        width: 160px;
-    }
-}
-
-.mc-requirements-requirement-view {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-}
-
-.mc-requirements-container {
-    display: flex;
-    // filter: drop-shadow(2px 4px 6px black);
-    flex-direction: row;
-    justify-content: space-around;
-    // max-width: calc(1920px - 2 * ${theme.spacing(12)});
-    padding: 160px 0;
-    // width: 100%;
-
-    & h5 {
-        // align: center;
-        color: rgb(255, 255, 255);
-        font-family: Pangram;
-        font-size: 32px;
-        font-weight: 700;
-        letter-spacing: 0.4px;
-        line-height: 40px;
-    }
-
-    & h2 {
-        font-family: "Balsamiq Sans",Roboto,Arial;
-    }
-}
