--- assets/public/assets/styles/ui/home/stripe.module.scss
@@ -1,19 +1,0 @@
-@use "@styles/variables.module.scss";
-
-.mc-stripe-text {
-    color: variables.$black;
-    font-family: "Monument Extended";
-    font-size: 64px;
-    font-weight: 400;
-    letter-spacing: 4px;
-    line-height: 76.8px;
-    text-transform: uppercase;
-}
-
-.mc-stripe-container {
-    background-color: variables.$yellow;
-    display: flex;
-    justify-content: space-around;
-    padding: 24px 0;
-    width: 100%;
-}
